Web UI & Flask Integration/requirements.txt
asn1crypto==1.4.0
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
Flask==2.0.2
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
mysqlclient==2.0.3
numpy==1.19.5
oscrypto==1.2.1
pandas==1.1.5
protobuf==3.19.1
pycparser==2.20
pycryptodomex==3.11.0
PyJWT==2.3.0
PyMySQL==1.0.2
pyOpenSSL==20.0.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
wincertstore==0.2
zipp==3.6.0
